# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evidentia API",
    "description": "Evidence platform for field media: signed Cloudinary ingestion, AI enrichment, verification and traceable reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/before-after/suggest": {
      "post": {
        "tags": [
          "before-after"
        ],
        "summary": "Suggest",
        "description": "Find before/after candidates: same site, PostGIS distance, baseline vs endline windows,\nranked by SigLIP similarity. Each new candidate is queued for geometric verification, alignment\nand change detection; the ranking is final once those finish.",
        "operationId": "suggest_v1_projects__project_id__before_after_suggest_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/before-after": {
      "get": {
        "tags": [
          "before-after"
        ],
        "summary": "List Pairs",
        "operationId": "list_pairs_v1_projects__project_id__before_after_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PairStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PairSummary"
                  },
                  "title": "Response List Pairs V1 Projects  Project Id  Before After Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "before-after"
        ],
        "summary": "Create Pair",
        "description": "Compare two specific photos (queued for analysis like a suggested pair).",
        "operationId": "create_pair_v1_projects__project_id__before_after_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/before-after/{pair_id}": {
      "get": {
        "tags": [
          "before-after"
        ],
        "summary": "Get Pair",
        "operationId": "get_pair_v1_before_after__pair_id__get",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pair Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/before-after/{pair_id}/analyze": {
      "post": {
        "tags": [
          "before-after"
        ],
        "summary": "Analyze",
        "description": "Re-run alignment and change detection (history of decisions is kept).",
        "operationId": "analyze_v1_before_after__pair_id__analyze_post",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pair Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/before-after/{pair_id}/confirm": {
      "post": {
        "tags": [
          "before-after"
        ],
        "summary": "Confirm",
        "description": "A reviewer confirms the photos show the same place: the pair becomes usable as evidence.",
        "operationId": "confirm_v1_before_after__pair_id__confirm_post",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pair Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/before-after/{pair_id}/reject": {
      "post": {
        "tags": [
          "before-after"
        ],
        "summary": "Reject",
        "operationId": "reject_v1_before_after__pair_id__reject_post",
        "parameters": [
          {
            "name": "pair_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pair Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PairCreate": {
        "properties": {
          "before_asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Before Asset Id"
          },
          "after_asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "After Asset Id"
          }
        },
        "type": "object",
        "required": [
          "before_asset_id",
          "after_asset_id"
        ],
        "title": "PairCreate"
      },
      "PairDecision": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "PairDecision"
      },
      "PairDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "status": {
            "$ref": "#/components/schemas/PairStatus"
          },
          "origin": {
            "$ref": "#/components/schemas/PairOrigin"
          },
          "before": {
            "$ref": "#/components/schemas/PairPhoto"
          },
          "after": {
            "$ref": "#/components/schemas/PairPhoto"
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "days_apart": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Apart"
          },
          "rank_score": {
            "type": "number",
            "title": "Rank Score"
          },
          "alignment_grade": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "strong",
                  "partial",
                  "none"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment Grade",
            "description": "None until analysed"
          },
          "change_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Summary"
          },
          "changed_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed Share"
          },
          "green_delta_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Green Delta Pp"
          },
          "limitation_count": {
            "type": "integer",
            "title": "Limitation Count"
          },
          "analysis_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Error"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "alignment": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "changes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changes"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "analysis_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Version"
          },
          "analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          },
          "before_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Url",
            "description": "Signed review rendition of the before photo"
          },
          "after_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Url"
          },
          "composite_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composite Url",
            "description": "Signed Cloudinary side-by-side composite"
          },
          "composite_transformation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composite Transformation"
          },
          "public_composite_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Composite Url",
            "description": "Face-pixelated composite (external use)"
          },
          "public_composite_transformation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Composite Transformation"
          },
          "created_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id"
          },
          "decided_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Id"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/PairClaimRef"
            },
            "type": "array",
            "title": "Claims"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "site_id",
          "status",
          "origin",
          "before",
          "after",
          "similarity",
          "distance_m",
          "days_apart",
          "rank_score",
          "alignment_grade",
          "change_summary",
          "changed_share",
          "green_delta_pp",
          "limitation_count",
          "analysis_error",
          "decided_at",
          "created_at",
          "alignment",
          "changes",
          "limitations",
          "analysis_version",
          "analyzed_at",
          "before_url",
          "after_url",
          "composite_url",
          "composite_transformation",
          "public_composite_url",
          "public_composite_transformation",
          "created_by_id",
          "decided_by_id",
          "decision_note"
        ],
        "title": "PairDetail"
      },
...
      "PairStatus": {
        "type": "string",
        "enum": [
          "candidate",
          "analyzing",
          "analyzed",
          "failed",
          "confirmed",
          "rejected"
        ],
        "title": "PairStatus"
      },
      "PairSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "status": {
            "$ref": "#/components/schemas/PairStatus"
          },
          "origin": {
            "$ref": "#/components/schemas/PairOrigin"
          },
          "before": {
            "$ref": "#/components/schemas/PairPhoto"
          },
          "after": {
            "$ref": "#/components/schemas/PairPhoto"
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "days_apart": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Apart"
          },
          "rank_score": {
            "type": "number",
            "title": "Rank Score"
          },
          "alignment_grade": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "strong",
                  "partial",
                  "none"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment Grade",
            "description": "None until analysed"
          },
          "change_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Summary"
          },
          "changed_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed Share"
          },
          "green_delta_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Green Delta Pp"
          },
          "limitation_count": {
            "type": "integer",
            "title": "Limitation Count"
          },
          "analysis_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Error"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "site_id",
          "status",
          "origin",
          "before",
          "after",
          "similarity",
          "distance_m",
          "days_apart",
          "rank_score",
          "alignment_grade",
          "change_summary",
          "changed_share",
          "green_delta_pp",
          "limitation_count",
          "analysis_error",
          "decided_at",
          "created_at"
        ],
        "title": "PairSummary"
      },
...
      "SuggestOut": {
        "properties": {
          "windows": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Windows"
              },
              {
                "type": "null"
              }
            ]
          },
          "anchors": {
            "type": "integer",
            "title": "Anchors",
            "description": "After photos that found at least one candidate"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "queued": {
            "type": "integer",
            "title": "Queued"
          },
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/PairSummary"
            },
            "type": "array",
            "title": "Pairs"
          }
        },
        "type": "object",
        "required": [
          "windows",
          "anchors",
          "created",
          "queued",
          "pairs"
        ],
        "title": "SuggestOut"
      },
      "SuggestRequest": {
        "properties": {
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "baseline_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline From"
          },
          "baseline_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline To"
          },
          "endline_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endline From"
          },
          "endline_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endline To"
          },
          "anchor_asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Asset Id",
            "description": "Find 'before' photos for this 'after' photo only"
          },
          "max_distance_m": {
            "type": "number",
            "maximum": 5000.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Distance M",
            "default": 250.0
          },
          "min_gap_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Min Gap Days",
            "default": 7
          },
          "per_anchor": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Per Anchor",
            "default": 3
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 60
          }
        },
        "type": "object",
        "title": "SuggestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Windows": {
        "properties": {
          "baseline_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline From"
          },
          "baseline_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline To"
          },
          "endline_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endline From"
          },
          "endline_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endline To"
          }
        },
        "type": "object",
        "required": [
          "baseline_from",
          "baseline_to",
          "endline_from",
          "endline_to"
        ],
        "title": "Windows"
      }
    }
  }
}